Using new config validation

+++ src/test/java/io/confluent/connect/hdfs/FailureRecoveryTest.java
@@ -70,8 +70,7 @@ public class FailureRecoveryTest extends HdfsSinkConnectorTestBase {
     storage.setFailure(MemoryStorage.Failure.appendFailure);
 
     hdfsWriter.write(sinkRecords);
-    assertEquals(context.timeout(), connectorConfig.getLong(
-        HdfsSinkConnectorConfig.RETRY_BACKOFF_CONFIG));
+    assertEquals((long) context.timeout(), (long) connectorConfig.getLong(HdfsSinkConnectorConfig.RETRY_BACKOFF_CONFIG));
 
     Map<String, List<Object>> data = Data.getData();
 
@@ -126,7 +125,7 @@ public class FailureRecoveryTest extends HdfsSinkConnectorTestBase {
     writer.setFailure(MemoryRecordWriter.Failure.writeFailure);
     hdfsWriter.write(sinkRecords);
 
-    assertEquals(context.timeout(), connectorConfig.getLong(HdfsSinkConnectorConfig.RETRY_BACKOFF_CONFIG));
+    assertEquals((long) context.timeout(), (long) connectorConfig.getLong(HdfsSinkConnectorConfig.RETRY_BACKOFF_CONFIG));
 
     Map<String, List<Object>> data = Data.getData();
     String directory2 = TOPIC + "/" + "partition=" + String.valueOf(PARTITION2);
@@ -143,7 +142,7 @@ public class FailureRecoveryTest extends HdfsSinkConnectorTestBase {
 
     writer.setFailure(MemoryRecordWriter.Failure.closeFailure);
     hdfsWriter.write(new ArrayList<SinkRecord>());
-    assertEquals(context.timeout(), connectorConfig.getLong(HdfsSinkConnectorConfig.RETRY_BACKOFF_CONFIG));
+    assertEquals((long) context.timeout(), (long) connectorConfig.getLong(HdfsSinkConnectorConfig.RETRY_BACKOFF_CONFIG));
 
     Map<String, String> tempFileNames = hdfsWriter.getTempFileNames(TOPIC_PARTITION);
     String tempFileName = tempFileNames.get(encodedPartition);
@@ -194,7 +193,7 @@ public class FailureRecoveryTest extends HdfsSinkConnectorTestBase {
 
     writer.setFailure(MemoryRecordWriter.Failure.writeFailure);
     hdfsWriter.write(sinkRecords);
-    assertEquals(context.timeout(), connectorConfig.getLong(HdfsSinkConnectorConfig.RETRY_BACKOFF_CONFIG));
+    assertEquals((long) context.timeout(), (long) connectorConfig.getLong(HdfsSinkConnectorConfig.RETRY_BACKOFF_CONFIG));
 
     writer.setFailure(MemoryRecordWriter.Failure.closeFailure);
     // nothing happens as we the retry back off hasn't yet passed
